delete help_info.txt,help_start in the test references and clean code.

+++ calvalus-processing/src/main/java/com/bc/calvalus/processing/prevue/PrevueMapper.java
@@ -152,8 +152,8 @@ public class PrevueMapper extends Mapper<NullWritable, NullWritable, NullWritabl
         reprojectionOp.setParameter("pixelSizeY", 1.0 / 112.0);
         reprojectionOp.setParameter("width", 49);
         reprojectionOp.setParameter("height", 49);
-        reprojectionOp.setParameter("easting", (double) location.getLon());
-        reprojectionOp.setParameter("northing", (double) location.getLat());
+        reprojectionOp.setParameter("easting", location.getLon());
+        reprojectionOp.setParameter("northing", location.getLat());
         return reprojectionOp;
     }
